pyproject.toml
[tool.poetry]
name = "coinflow"
version = "2.0.0"
description = "Advanced Telegram Bot for currency conversion and cryptocurrency analysis"
authors = ["bobberdolle1 <kirillkrovyakov@yandex.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32"
python-telegram-bot = "^22.5"
python-dotenv = "^1.0"
matplotlib = "^3.9"
yfinance = "^0.2"
numpy = "^2.0"
statsmodels = "^0.14"
scikit-learn = "^1.5"
apscheduler = "^3.10"
bestchange-api = "^3.1"
sqlalchemy = "^2.0"
aiohttp = "^3.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"